Refactoring everything following Python and MetaWards update (incomplete)
