--- tools/lockdown/iterate.py
@@ -1,67 +1,0 @@
-# Moved imports to local space
-
-
-def get_lockdown_state(network, population):
-    from datetime import datetime
-
-    date = population.date
-    params = network.params.user_params
-
-    y1 = int(params["lockdown_date_1_year"])
-    m1 = int(params["lockdown_date_1_month"])
-    d1 = int(params["lockdown_date_1_day"])
-
-    y2 = int(params["lockdown_date_2_year"])
-    m2 = int(params["lockdown_date_2_month"])
-    d2 = int(params["lockdown_date_2_day"])
-
-    # Lock down dates
-    lock_1 = datetime(y1, m1, d1).date()
-    lock_2 = datetime(y2, m2, d2).date()
-    
-    state = 0
-    rate = 1.0
-    if date >= lock_1:
-        state += 1
-        rate = params["scale_rate"][1]
-    if date >= lock_2:
-        state += 1
-        rate = (1.0 - ((1.0 - params["scale_rate"][1]) * params["scale_rate"][2]))
-
-    return state, rate
-
-def advance_lockdown(network, population, **kwargs):
-    from metawards.iterators import advance_infprob
-    from metawards.iterators import advance_play
-    from metawards.iterators import advance_fixed
-
-    params = network.params
-    state, rate = get_lockdown_state(network, population)
-    can_work = params.user_params["can_work"][state]
-    print(f"state {state}: scale_rate = {rate}, can_work = {can_work}")
-
-    advance_infprob(scale_rate=rate,
-                    network=network, population=population,
-                    **kwargs)
-    advance_play(network=network, population=population, **kwargs)
-
-    if can_work:
-        advance_fixed(network=network, population=population, **kwargs)
-
-def iterate_custom(network, population, **kwargs):
-    from metawards.iterators import iterate_working_week
-
-    params = network.params
-    
-    state, rate = get_lockdown_state(network, population)
-    can_work = params.user_params["can_work"][state]
-    
-    print("Simulation date " + str(population.date))
-    if state > 0:
-        return [advance_lockdown]
-    else:
-        print(f"state {state}: scale_rate = {rate}, can_work = {can_work}")
-        print("Normal working week day")
-        return iterate_working_week(network=network,
-                                    population=population,
-                                    **kwargs)
